新增PLIC驱动程序

+++ bsp/lib/perip/include/spi.h
@@ -18,7 +18,7 @@
 #define SPI0 SPI0_BASE
 
 
-void spi_cp_model(uint32_t SPIx, uint32_t spi_cpmodel);//SPI相位控制
+void spi_cp_model(uint32_t SPIx, uint8_t spi_cpmodel);//SPI相位控制
 void spi_sclk_div(uint32_t SPIx, uint32_t spi_div);//SPI SCLK分频器配置
 void spi_set_cs(uint32_t SPIx, uint32_t spi_cs);//SPI CS片选信号控制
 void spi_send_byte(uint32_t SPIx, uint32_t data);//SPI发送字节，发完不管
